First working version of git-server using embedded dropbear on port 222, gitolite with gitweb integration. Install wizards prompts for gitolite admin public key. Needs clean up inn dsm-control and install script. No git over http(s) support yet.

+++ spk/git-server/src/dsm-control.sh
@@ -12,12 +12,13 @@ DROPBEAR_PID_FILE="${INSTALL_DIR}/var/run/dropbear.pid"
 PID_FILE="${INSTALL_DIR}/var/run/git-daemon.pid"
 LOG_FILE="${INSTALL_DIR}/var/log/git-daemon.log"
 GIT_HOME="${INSTALL_DIR}/var/home"
-BASE_PATH="${INSTALL_DIR}/var/repositories"
+BASE_PATH="${INSTALL_DIR}/var/home/repositories"
 
 start_daemon ()
 {
-    su - ${USER} -c "${GIT_DIR}/bin/git daemon --export-all --base-path=${BASE_PATH} --pid-file=${PID_FILE} --reuseaddr --verbose --detach ${BASE_PATH}"
-    su - ${USER} -c "LD_LIBRARY_PATH=${INSTALL_DIR}/lib ${INSTALL_DIR}/sbin/dropbear -w"
+    su - ${USER} -c "${GIT_DIR}/bin/git daemon --base-path=${BASE_PATH} --pid-file=${PID_FILE} --reuseaddr --verbose --detach ${BASE_PATH}"
+    export LD_LIBRARY_PATH=${INSTALL_DIR}/lib
+    ${INSTALL_DIR}/sbin/dropbear -w
 
     # Symlink for gitweb
     ln -s ${GIT_DIR}/share/gitweb /var/services/web/gitweb
@@ -32,25 +33,29 @@ start_daemon ()
 stop_daemon ()
 {
     kill `cat ${PID_FILE}`
-    wait_for_status 1 20 || kill -9 `cat ${PID_FILE}`
-    rm -f ${PID_FILE}
-
     kill `cat ${DROPBEAR_PID_FILE}`
+    rm -f ${PID_FILE}
     rm -f ${DROPBEAR_PID_FILE}
 
-
+    # Unlink gitweb
     rm -f /var/services/web/gitweb
     rm -f /usr/syno/etc/sites-enabled-user/gitweb.apache2.conf
+
     # Restart Apache
     /usr/syno/etc/rc.d/S97apache-user.sh restart
 }
 
 daemon_status ()
 {
-    if [ -f ${PID_FILE} ] && kill -0 `cat ${PID_FILE}` > /dev/null 2>&1; then
+    if [ -f ${PID_FILE} ] && kill -0 `cat ${PID_FILE}` && [ -f ${DROPBEAR_PID_FILE} ] && kill -0 `cat ${DROPBEAR_PID_FILE}` > /dev/null 2>&1; then
         return
     fi
-    rm -f ${PID_FILE}
+    if ! `pidof git-daemon`; then
+        rm -f ${PID_FILE}
+    fi
+    if ! `pidof dropbear`; then
+        rm -f ${DROPBEAR_PID_FILE}
+    fi
     return 1
 }
